src/test/java/com/example/kn/service/CustomerServiceTest.java
package com.example.kn.service;

import com.example.kn.model.Customer;
import com.example.kn.repository.CustomerRepository;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.when;

@ExtendWith(MockitoExtension.class)
public class CustomerServiceTest {

    @InjectMocks
    private CustomerService customerService;

    @Mock
    private CustomerRepository customerRepository;

    @Test
    public void testCreateCustomer() {

        Customer customer = new Customer();
        customer.setFullName("John Doe");
        customer.setRegistrationCode("REG123");
        customer.setEmail("john@example.com");
        customer.setTelephone("1234567890");

        when(customerRepository.save(any())).thenReturn(customer);

        Customer savedCustomer = customerService.createCustomer(customer);

        assertNotNull(savedCustomer);
        assertEquals("John Doe", savedCustomer.getFullName());

    }


}
